feat: constructor
