+++ src/components/constructor/index.ts
@@ -1,0 +1,1 @@
+export { Constructor } from './constructor'
